Create README.md
This code has been developed by the author in order to perform the fractal analysis of particle contour.It characterises the morphology of soil particles over the range of experimentally accessible scalesthrough three new non-dimensional quantitative morphological descriptors characterising:  (i) overallparticle shape at the macro-scale, (ii) particle regularity at the meso-scale, and (iii) particle texture atthe micro-scale.  The characteristic size separating structural features and textural features emergesdirectly from the results of the fractal analysis of the contour of the particle, and is a decreasingfraction of particle dimension.